adds missing dependency, fixes imports

+++ test/echo/HelperContract.sol
@@ -1,27 +1,25 @@
 // SPDX-License-Identifier: UNLICENSED
 pragma solidity 0.8.18;
 
-/******************************************************************************\
-* Authors: Timo Neumann <timo@fyde.fi>
-* EIP-2535 Diamonds: https://eips.ethereum.org/EIPS/eip-2535
-* Helper functions for the translation from the jest tests in the original repo
-* to solidity tests.
-/******************************************************************************/
+/**
+ * \
+ * Authors: Timo Neumann <timo@fyde.fi>
+ * EIP-2535 Diamonds: https://eips.ethereum.org/EIPS/eip-2535
+ * Helper functions for the translation from the jest tests in the original repo
+ * to solidity tests.
+ * /*****************************************************************************
+ */
 
-import "../../lib/solidity-stringutils/strings.sol";
+import "solidity-stringutils/strings.sol";
 import "../../src/echo/contracts/interfaces/IDiamond.sol";
 import "../../src/echo/contracts/interfaces/IDiamondLoupe.sol";
-import "../../lib/forge-std/src/Test.sol";
+import "forge-std/Test.sol";
 
-
-abstract contract HelperContract is IDiamond, IDiamondLoupe, Test{
+abstract contract HelperContract is IDiamond, IDiamondLoupe, Test {
     using strings for *;
 
     // return array of function selectors for given facet name
-    function generateSelectors(string memory _facetName)
-    internal
-    returns (bytes4[] memory selectors)
-    {
+    function generateSelectors(string memory _facetName) internal returns (bytes4[] memory selectors) {
         //get string of contract methods
         string[] memory cmd = new string[](4);
         cmd[0] = "forge";
@@ -35,9 +33,9 @@ abstract contract HelperContract is IDiamond, IDiamondLoupe, Test{
         strings.slice memory s = st.toSlice();
 
         // Skip TRACE lines if any
-        strings.slice memory nl = '\n'.toSlice();
-        strings.slice memory trace = 'TRACE'.toSlice();
-        while (s.contains(trace) ) {
+        strings.slice memory nl = "\n".toSlice();
+        strings.slice memory trace = "TRACE".toSlice();
+        while (s.contains(trace)) {
             s.split(nl);
         }
 
@@ -46,48 +44,46 @@ abstract contract HelperContract is IDiamond, IDiamondLoupe, Test{
         strings.slice memory dbquote = '"'.toSlice();
         selectors = new bytes4[]((s.count(colon)));
 
-        for(uint i = 0; i < selectors.length; i++) {
-            s.split(dbquote);   // advance to next doublequote
+        for (uint256 i = 0; i < selectors.length; i++) {
+            s.split(dbquote); // advance to next doublequote
             // split at colon, extract string up to next doublequote for methodname
             strings.slice memory method = s.split(colon).until(dbquote);
             selectors[i] = bytes4(method.keccak());
-            strings.slice memory selectr = s.split(comma).until(dbquote);     // advance s to the next comma
+            strings.slice memory selectr = s.split(comma).until(dbquote); // advance s to the next comma
         }
         return selectors;
     }
 
-// helper to remove index from bytes4[] array
-    function removeElement(uint index, bytes4[] memory array) public pure returns (bytes4[] memory){
+    // helper to remove index from bytes4[] array
+    function removeElement(uint256 index, bytes4[] memory array) public pure returns (bytes4[] memory) {
         bytes4[] memory newarray = new bytes4[](array.length-1);
-        uint j = 0;
-        for(uint i = 0; i < array.length; i++){
-            if (i != index){
+        uint256 j = 0;
+        for (uint256 i = 0; i < array.length; i++) {
+            if (i != index) {
                 newarray[j] = array[i];
                 j += 1;
             }
         }
         return newarray;
-
     }
 
-// helper to remove value from bytes4[] array
-    function removeElement(bytes4 el, bytes4[] memory array) public pure returns (bytes4[] memory){
-        for(uint i = 0; i < array.length; i++){
-            if (array[i] == el){
+    // helper to remove value from bytes4[] array
+    function removeElement(bytes4 el, bytes4[] memory array) public pure returns (bytes4[] memory) {
+        for (uint256 i = 0; i < array.length; i++) {
+            if (array[i] == el) {
                 return removeElement(i, array);
             }
         }
         return array;
-
     }
 
     function removeElements(bytes4[] memory elements, bytes4[] memory array) public pure returns (bytes4[] memory) {
         bytes4[] memory newArray = new bytes4[](array.length);
-        uint count = 0;
+        uint256 count = 0;
 
-        for (uint i = 0; i < array.length; i++) {
+        for (uint256 i = 0; i < array.length; i++) {
             bool toRemove = false;
-            for (uint j = 0; j < elements.length; j++) {
+            for (uint256 j = 0; j < elements.length; j++) {
                 // console.log('comparing two elements');
                 // console.logBytes4( array[i]);
                 // console.logBytes4( elements[j]);
@@ -104,7 +100,7 @@ abstract contract HelperContract is IDiamond, IDiamondLoupe, Test{
         }
 
         bytes4[] memory resultArray = new bytes4[](count);
-        for (uint i = 0; i < count; i++) {
+        for (uint256 i = 0; i < count; i++) {
             resultArray[i] = newArray[i];
         }
 
@@ -112,7 +108,7 @@ abstract contract HelperContract is IDiamond, IDiamondLoupe, Test{
     }
 
     function containsElement(bytes4[] memory array, bytes4 el) public pure returns (bool) {
-        for (uint i = 0; i < array.length; i++) {
+        for (uint256 i = 0; i < array.length; i++) {
             if (array[i] == el) {
                 return true;
             }
@@ -122,7 +118,7 @@ abstract contract HelperContract is IDiamond, IDiamondLoupe, Test{
     }
 
     function containsElement(address[] memory array, address el) public pure returns (bool) {
-        for (uint i = 0; i < array.length; i++) {
+        for (uint256 i = 0; i < array.length; i++) {
             if (array[i] == el) {
                 return true;
             }
@@ -135,8 +131,8 @@ abstract contract HelperContract is IDiamond, IDiamondLoupe, Test{
         if (array1.length != array2.length) {
             return false;
         }
-        for (uint i = 0; i < array1.length; i++) {
-            if (containsElement(array1, array2[i])){
+        for (uint256 i = 0; i < array1.length; i++) {
+            if (containsElement(array1, array2[i])) {
                 return true;
             }
         }
@@ -144,18 +140,18 @@ abstract contract HelperContract is IDiamond, IDiamondLoupe, Test{
         return false;
     }
 
-    function getAllSelectors(address diamondAddress) public view returns (bytes4[] memory){
+    function getAllSelectors(address diamondAddress) public view returns (bytes4[] memory) {
         Facet[] memory facetList = IDiamondLoupe(diamondAddress).facets();
 
-        uint len = 0;
-        for (uint i = 0; i < facetList.length; i++) {
+        uint256 len = 0;
+        for (uint256 i = 0; i < facetList.length; i++) {
             len += facetList[i].functionSelectors.length;
         }
 
-        uint pos = 0;
+        uint256 pos = 0;
         bytes4[] memory selectors = new bytes4[](len);
-        for (uint i = 0; i < facetList.length; i++) {
-            for (uint j = 0; j < facetList[i].functionSelectors.length; j++) {
+        for (uint256 i = 0; i < facetList.length; i++) {
+            for (uint256 j = 0; j < facetList[i].functionSelectors.length; j++) {
                 selectors[pos] = facetList[i].functionSelectors[j];
                 pos += 1;
             }
@@ -170,17 +166,11 @@ abstract contract HelperContract is IDiamond, IDiamondLoupe, Test{
     function facetFunctionSelectors(address _facet) external view returns (bytes4[] memory facetFunctionSelectors_) {}
     function facets() external view returns (Facet[] memory facets_) {}
 
-
-    function getEthSignedMessageHash(
-        bytes32 _messageHash
-    ) public pure returns (bytes32) {
+    function getEthSignedMessageHash(bytes32 _messageHash) public pure returns (bytes32) {
         /*
         Signature is produced by signing a keccak256 hash with the following format:
         "\x19Ethereum Signed Message\n" + len(msg) + msg
         */
-        return
-            keccak256(
-                abi.encodePacked("\x19Ethereum Signed Message:\n32", _messageHash)
-            );
+        return keccak256(abi.encodePacked("\x19Ethereum Signed Message:\n32", _messageHash));
     }
 }
